+++ lirc/__init__.py
@@ -1,6 +1,6 @@
 from lirc.client import Client
 from lirc.connection.lircd_connection import LircdConnection
 
-__version__ = "0.1.0"
+__version__ = "0.2.0"
 
 __all__ = ["Client", "LircdConnection"]