+++ CHANGELOG.rst
@@ -8,7 +8,7 @@ Changelog <https://keepachangelog.com/en/1.0.0/>`_, and this project
 adheres to `Semantic Versioning <https://semver.org/spec/v2.0.0.html>`_.
 
 
-Unreleased
+0.2.0 - 2020-12-13
 ------------------
 
 **Added**